Fixed issue #93 Use built in zip libraries to extract elastic search.
I also added a check to ensure that JAVA_HOME has been configured.

+++ Libraries/Start-ElasticSearch.ps1
@@ -3,14 +3,20 @@ if ($env:ES_VERSION) {
     $es_version = $env:ES_VERSION
 }
 
+if (!(Test-Path -Path $env:JAVA_HOME)) {
+    Write-Error "Please ensure the latest version of java is installed and the JAVA_HOME environmental variable has been set."
+    Return
+}
+
 Push-Location $PSScriptRoot
 
 if (!(Test-Path -Path "elasticsearch-$es_version") -And !(Test-Path -Path "elasticsearch.zip")) {
     wget "http://download.elasticsearch.org/elasticsearch/elasticsearch/elasticsearch-$es_version.zip" -OutFile "elasticsearch.zip"
 }
 
-if (Test-Path -Path "elasticsearch.zip") {
-    7z x -y "elasticsearch.zip" > $null
+if ((Test-Path -Path "elasticsearch.zip") -And !(Test-Path -Path "elasticsearch-$es_version")) {
+    Add-Type -assembly "system.io.compression.filesystem"
+    [io.compression.zipfile]::ExtractToDirectory("$PSScriptRoot\elasticsearch.zip", $PSScriptRoot)
     cp .\elasticsearch.yml .\elasticsearch-$es_version\config -Force
     rm elasticsearch.zip
 }
